% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_all_macroarea}
\alias{get_all_macroarea}
\title{Get all the macroareas' name}
\usage{
get_all_macroarea(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{macroareas} names for the object \code{x}.
}
\description{
This is a generic function aimed to access to the \code{macroareas}
names of an object
}
