% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_all_macroregion}
\alias{get_all_macroregion}
\title{Get all the macroregions' names}
\usage{
get_all_macroregion(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{macroregions}' name for the object \code{x}.
}
\description{
This is a generic function aimed to access to the \code{macroregions}
names of an object
}
