% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_all_region}
\alias{get_all_region}
\title{Get the all included regions' name}
\usage{
get_all_region(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The regions' name for the object \code{x}.
}
\description{
This is a generic function aimed to access to the all the regions' name
into an object
}
