% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_center}
\alias{get_center}
\title{Get the centers name}
\usage{
get_center(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{center} name for the object \code{x}.
}
\description{
This is a generic function aimed to access to the \code{center} name of a
center.
}
