% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_centers}
\alias{get_centers}
\title{Get the centers details}
\usage{
get_centers(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{centers} details for the object \code{x}.
}
\description{
This is a generic function aimed to access to the \code{centers} details
of an object.
}
