% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_current_strip}
\alias{get_current_strip}
\title{Get the current strip details}
\usage{
get_current_strip(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{current_strip} details for the object \code{x}.
}
\description{
This is a generic function aimed to access to the \code{current_strip}
details of an object.
}
