% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_final_p_macroregion_by_time.R
\name{get_final_p_macroregion_by_time}
\alias{get_final_p_macroregion_by_time}
\title{Final probability distribution of region in a macroregion}
\usage{
get_final_p_macroregion_by_time(mr, n = 1L, state)
}
\arguments{
\item{mr}{an object of class \code{macroregion}}

\item{n}{number of stages to compute (default is 1)}

\item{state}{an object of class \code{state}}
}
\value{
a list of dataframes, each element of the list is a time (or an ID
        for every further organ), each dataframe report the final
}
\description{
Final probability distribution of region in a macroregion
}
