% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_macroareas}
\alias{get_macroareas}
\title{Get the macroareas details}
\usage{
get_macroareas(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{macroareas} details for the object \code{x}.
}
\description{
This is a generic function aimed to access to the \code{macroareas}
details of an object
}
