% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_macroregions}
\alias{get_macroregions}
\title{Get the macroregions details}
\usage{
get_macroregions(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{macroregions} details for the object \code{x}.
}
\description{
This is a generic function aimed to access to the \code{macroregions}
details of an object
}
