% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_offered}
\alias{get_offered}
\title{Get number of offered organs}
\usage{
get_offered(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The number of offered organs for the object \code{x}.
}
\description{
This is a generic function aimed to access (or compute) to the
number of offered organs of a center, region, ...
}
