% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_p_accept}
\alias{get_p_accept}
\title{Get the probability of acceptance}
\usage{
get_p_accept(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The probability of acceptance for the object \code{x}.
}
\description{
This is a generic function aimed to access (or compute) to the
probability of acceptance (for an offered organ) details of an object
}
