% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_accept_by_position.R
\name{get_p_accept_by_position}
\alias{get_p_accept_by_position}
\title{Acceptance matrices by macroregions' position}
\usage{
get_p_accept_by_position(state)
}
\arguments{
\item{state}{an object of class \code{state}}
}
\value{
a list of n matrix where n is the number of macroareas.
}
\description{
In the case of the organ is offered to a macroarea, which are the
probability that a (macro)reagion i obtains (accepting...) it if it is in
position j?
}
