% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_accept_by_position_nitp.R
\name{get_p_accept_by_position_nitp}
\alias{get_p_accept_by_position_nitp}
\title{Acceptance matrices for macroregions' position}
\usage{
get_p_accept_by_position_nitp(macroregion, state)
}
\arguments{
\item{macroregion}{an object of class \code{macroregion}}

\item{state}{an object of class \code{stae}}
}
\value{
a matrix
}
\description{
In the case of the organ is offered to a macroregion, which are the
probability its reagions to obtains (accepting...) it if it is in
position j?
}
