% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_p_default}
\alias{get_p_default}
\title{Get the default probability of acceptance for a region}
\usage{
get_p_default(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The probability of acceptance for the object \code{x}.
}
\description{
This is a generic function aimed to access to the default probability
attributes of acceptance (for an offered organ) of an object
}
