% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_macroareas.R
\name{get_p_macroareas}
\alias{get_p_macroareas}
\title{Macroareas' probability of usage}
\usage{
get_p_macroareas(state)
}
\arguments{
\item{state}{an object of class \code{state}.}
}
\value{
a \code{data.frame} with two columns, the first reporting the
        macroareas' name or "lost", and the second reporting the probability
        that a "next" organ will be effectively used by that macroarea (or
        will be lost).
}
\description{
this function comput the probabilities for macroareas to use a "next" organ
based on the portion of organs offered and the probability to accept them
from the centers. The probability that the organ will be lost is also
returned.
}
\examples{
\dontrun{
  get_p_macroareas(italy)
}
}
