% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_positions_by_time.R
\name{get_p_position_by_time}
\alias{get_p_position_by_time}
\title{positions' probability by stage}
\usage{
get_p_position_by_time(state, n = 1L)
}
\arguments{
\item{state}{an object of class \code{state}}

\item{n}{number of stages to compute (default is 1)}
}
\value{
a list of length n in which element is stored a list containing the
        macroareas' positions' probability matrix
}
\description{
This function comput the matrices of positions' probability for each stage
requested.
}
