% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_position_by_time_nitp.R
\name{get_p_position_by_time_nitp}
\alias{get_p_position_by_time_nitp}
\title{positions' probability by stage for a macroegion}
\usage{
get_p_position_by_time_nitp(mr, n = 1, state)
}
\arguments{
\item{mr}{an object of class \code{macroregion}}

\item{n}{number of stages to compute (default is 1)}

\item{state}{an object of class \code{state}}
}
\value{
a list of length n in which element are stored the
        regions' positions' probability matrix
}
\description{
This function computes the matrices of positions' probability for each stage
requested.
}
