% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region.R
\name{get_region}
\alias{get_region}
\title{Get the region details}
\usage{
get_region(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The region details for the object \code{x}.
}
\description{
This is a generic function aimed to access to the region details of an
object
}
