% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_regions}
\alias{get_regions}
\title{Get the regions details}
\usage{
get_regions(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{regions} details for the object \code{x}.
}
\description{
This is a generic function aimed to access to the \code{regions} details
of an object.
}
