% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_state}
\alias{get_state}
\title{Get the state details}
\usage{
get_state(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The state details for the object \code{x}.
}
\description{
This is a generic function aimed to access to the state details of an
object.
}
