% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_time}
\alias{get_time}
\title{Get the current time details}
\usage{
get_time(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{current_time} details for the object \code{x}.
}
\description{
This is a generic function aimed to access to the \code{current_time}
details of an object.
}
