% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_macroregion}
\alias{is_macroregion}
\title{Test for inherithage on `macroregion`}
\usage{
is_macroregion(x)
}
\arguments{
\item{x}{an object used to select a method.}
}
\value{
TRUE if the object \code{x} inherits from `macrorigion`, FASLE
        otherwise.
}
\description{
This function aimed to check if an object is of class `macroregion`
}
