% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-macroarea.R, R/macroarea.R
\name{get_macroregions.macroarea}
\alias{get_macroregions.macroarea}
\alias{get_regions.macroarea}
\alias{get_all_region.macroarea}
\alias{get_all_macroregion.macroarea}
\alias{get_state.macroarea}
\alias{get_centers.macroarea}
\alias{get_p_accept.macroarea}
\alias{get_offered.macroarea}
\alias{get_initial_strip.macroarea}
\alias{get_current_strip.macroarea}
\alias{get_initial_time.macroarea}
\alias{get_time.macroarea}
\alias{get_final_time.macroarea}
\alias{macroarea}
\alias{print.macroarea}
\title{Constructor for object of class \code{macroarea}}
\usage{
\method{get_macroregions}{macroarea}(x, ...)

\method{get_regions}{macroarea}(x, ...)

\method{get_all_region}{macroarea}(x, ...)

\method{get_all_macroregion}{macroarea}(x, ...)

\method{get_state}{macroarea}(x, ...)

\method{get_centers}{macroarea}(x, ...)

\method{get_p_accept}{macroarea}(x, ...)

\method{get_offered}{macroarea}(x, ...)

\method{get_initial_strip}{macroarea}(x, ...)

\method{get_current_strip}{macroarea}(x, ...)

\method{get_initial_time}{macroarea}(x, ...)

\method{get_time}{macroarea}(x, ...)

\method{get_final_time}{macroarea}(x, ...)

macroarea(name, macroregions,
  initial_strip = get_all_macroregion(macroregions), initial_time = 0L,
  final_time = Inf)

\method{print}{macroarea}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{name}{[chr] the name of the macroarea}

\item{macroregions}{[lst] a list of objects of class \code{\link{region}}
and/or \code{\link{macroregion}}, passed by
\code{\link{set_macroregions}()}. Note: all the
\code{\link{region}}s and \code{\link{macroregion}}s in a
\code{\link{macroarea}} must be of the same state. Moreover, all
\code{\link{region}}s (alone and into some
\code{\link{macroregion}}) must be different.}

\item{initial_strip}{[chr] a sequence of \code{\link{macroregion}}s'
names in the \code{\link{macroarea}} representing the order in
which organs started to be distributed in the
\code{\link{macroarea}} at the beginning of the period. All the
listed \code{\link{region}}s and/or \code{\link{macroregion}}s
have to be included also in the \code{macroregions} field.
Exculsions are permitted but repetitions are not admitted.
(default is the sequence in which \code{\link{region}}s and/or
\code{\link{macroregion}}s has been included in the argument
\code{macroregions})}

\item{initial_time}{[int] (default is `0`) the time in which the
\code{\link{macroarea}} is defined start to be used}

\item{final_time}{[int] (default is `Inf`, i.e. forever) the time in
which the macroarea finish to be used}
}
\value{
An object of class \code{\link{macroarea}}.
}
\description{
The function \code{\link{macroarea}()} is the constructor function for
the homonymous class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_macroregions}: wrapper function to access to the 'macroregion'
detail.

\item \code{get_regions}: wrapper function to access to all the reagions in
macroarea

\item \code{get_all_region}: wrapper function to access to the all the
names of reagions appear in the `set_macroregion`,
recursively.

\item \code{get_all_macroregion}: wrapper function to access to the all the names
of macroregions (regions included).

\item \code{get_state}: wrapper function to access to the name of the
state

\item \code{get_centers}: wrapper function to access to all the centers of
a macroarea

\item \code{get_p_accept}: compute the probability that at least one center
in any region of the macroarea accept an offered organ.

\item \code{get_offered}: wrapper function to access to the detail
"offered".

\item \code{get_initial_strip}: wrapper function to access to the detail
"initial_strip".

\item \code{get_current_strip}: wrapper function to access to the detail
"current_strip".

\item \code{get_initial_time}: wrapper function to access to the detail
"initial_time".

\item \code{get_time}: wrapper function to access to the detail
"current_time".

\item \code{get_final_time}: wrapper function to access to the detail
"final_time".

\item \code{print}: nice (and coloured, if supported) print method.
}}

\note{
Even if improbable, a \code{\link{macroarea}} can be composed by a
      single \code{\link{macroregion}} which can be composed by a single
      \code{\link{region}} which can be composed by a single
      \code{\link{center}}
}
\examples{
pavia   <- center('Pavia',   'Lombardia', offered = 5, p_accept = 0.6)
pavia

bergamo <- center('Bergamo', 'Lombardia', 8)
milano  <- center('Milano',  'Lombardia', 10, 0.8)

lombardia <- region(set_centers(pavia, bergamo, milano), default_p = 0.7)
lombardia

padova <- center('Padova', 'Veneto', 8, 0.7)
veneto <- region(set_centers(padova))

nitp <- macroregion('NITp', regions = set_regions(lombardia, veneto),
  initial_strip = c('lombardia', 'lombardia', 'veneto')
)
nitp

torino   <- center('Torino', 'Piemonte', 7, 0.6)
piemonte <- region(set_centers(torino))

nord <- macroarea('Macroarea Nord',
  macroregions = set_macroregions(piemonte, nitp)
)
get_macroregions(nord)
get_regions(nord)
get_all_region(nord)
get_all_macroregion(nord)
get_state(nord) # "italy"
get_centers(nord)
get_p_accept(nord)
get_offered(nord)
get_initial_strip(nord)
get_current_strip(nord)
get_initial_time(nord)
get_time(nord)
get_final_time(nord)
pavia   <- center('Pavia',   'Lombardia', offered = 5, p_accept = 0.6)
pavia

bergamo <- center('Bergamo', 'Lombardia', 8)
milano  <- center('Milano',  'Lombardia', 10, 0.8)

lombardia <- region(set_centers(pavia, bergamo, milano), default_p = 0.7)
lombardia

padova <- center('Padova', 'Veneto', 8, 0.7)
veneto <- region(set_centers(padova))

nitp <- macroregion('NITp', regions = set_regions(lombardia, veneto),
  initial_strip = c('lombardia', 'lombardia', 'veneto')
)
nitp

torino   <- center('Torino', 'Piemonte', 7, 0.6)
piemonte <- region(set_centers(torino))

nord <- macroarea('Nord',
  macroregions = set_macroregions(piemonte, nitp)
)
nord
}
