% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{National regions.}
\format{A list with 1 slot per nation (actualy only one for) with a
  character vector reporting corresponding regions' name in
  alphabetical order.

\describe{
  \item{italy}{Character vector of Italian regions' name in alphabetical
    order
  }
}}
\source{
\url{https://it.wikipedia.org/wiki/Regioni_d\%27Italia}
}
\usage{
regions
}
\description{
A list of all national' regions. Actually it contains only the italian
ones.
}
\keyword{datasets}
