% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{set_centers}
\alias{set_centers}
\alias{print.set_centers}
\title{List of centers creator}
\usage{
set_centers(...)

\method{print}{set_centers}(x, ...)
}
\arguments{
\item{...}{a set of \code{\link{center}}}

\item{x}{an object used to select a method.}
}
\value{
The list of \code{\link{center}}s provided.
}
\description{
Function used to prepare and check list of centers to pass to
\code{centers} as needed to create \code{\link{region}} objects.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: nice (and coloured, if supported) print method.
}}

\examples{
pavia  <- center('Pavia',  'Lombardia')
milano <- center('Milano', 'Lombardia')

set_centers(pavia, milano)
}
