% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-set_macroareas.R, R/state.R
\name{get_all_macroarea.set_macroareas}
\alias{get_all_macroarea.set_macroareas}
\alias{set_macroareas}
\alias{print.set_macroareas}
\title{List of macroarea creator}
\usage{
\method{get_all_macroarea}{set_macroareas}(x, ...)

set_macroareas(...)

\method{print}{set_macroareas}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{a set of \code{\link{macroarea}}s.}
}
\value{
The list of \code{\link{macroarea}}s provided.
}
\description{
Function used to prepare and check list of \code{\link{macroarea}}s to
pass to \code{macroarea} as needed to create \code{\link{state}}
objects.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_all_macroarea}: wrapper function to access to the all the
names of macroareas appear in the `set_macroarea`,
recursively.

\item \code{print}: nice (and coloured, if supported) print method.
}}

\examples{
pavia  <- center('Pavia',  'Lombardia')
milano <- center('Milano', 'Lombardia')
lombardia <- region(set_centers(pavia, milano))

padova <- center('Padova', 'Veneto', 8, 0.7)
veneto <- region(set_centers(padova))

nitp <- macroregion('NITp', set_regions(lombardia, veneto))

torino   <- center('Torino', 'Piemonte', 7, 0.6)
piemonte <- region(set_centers(torino))

nord <- macroarea('Macroarea Nord',
  macroregions = set_macroregions(piemonte, nitp)
)

roma   <- center('Roma', 'Lazio', 10, 0.9)
lazio  <- region(set_centers(roma))

sud <- macroarea('Macroarea Sud',
  macroregions = set_macroregions(lazio)
)

get_all_macroarea(set_macroareas(nord, sud))
pavia     <- center('Pavia',   'Lombardia', offered = 5, p_accept = 0.6)
bergamo   <- center('Bergamo', 'Lombardia', 8)
milano    <- center('Milano',  'Lombardia', 10, 0.8)
lombardia <- region(set_centers(pavia, bergamo, milano), default_p = 0.7)

padova <- center('Padova', 'Veneto', 8, 0.7)
veneto <- region(set_centers(padova))

nitp <- macroregion('NITp', regions = set_regions(lombardia, veneto),
  initial_strip = c('lombardia', 'lombardia', 'veneto')
)

torino   <- center('Torino', 'Piemonte', 7, 0.6)
piemonte <- region(set_centers(torino))

nord <- macroarea('Macroarea Nord',
  macroregions = set_macroregions(piemonte, nitp)
)

roma   <- center('Roma', 'Lazio', 10, 0.9)
lazio  <- region(set_centers(roma))
sud    <- macroarea('Macroarea Sud',
  macroregions = set_macroregions(lazio)
)

set_macroareas(nord, sud)
}
