% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroarea.R
\name{set_macroregions}
\alias{set_macroregions}
\title{List of regions/macroregion creator}
\usage{
set_macroregions(...)
}
\arguments{
\item{...}{a set of \code{\link{region}}s and/or
\code{\link{macroregion}}}
}
\value{
The list of \code{\link{region}}s and \code{\link{macroregion}}s
        provided.
}
\description{
Function used to prepare and check list of \code{\link{region}}s and/or
\code{\link{macroregion}}s to pass to \code{macroregions} as needed to
create \code{\link{macroarea}} objects.
}
\examples{
pavia  <- center('Pavia',  'Lombardia')
milano <- center('Milano', 'Lombardia')
lombardia <- region(set_centers(pavia, milano))

padova <- center('Padova', 'Veneto', 8, 0.7)
veneto <- region(set_centers(padova))

nitp <- macroregion(set_regions(lombardia, veneto))

torino   <- center('Torino', 'Piemonte', 7, 0.6)
piemonte <- region(set_centers(torino))

set_macroregions(piemonte, nitp)
}
