% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_position.R
\name{transition_position}
\alias{transition_position}
\title{Transition of positions for a next organ}
\usage{
transition_position(Tpast, p)
}
\arguments{
\item{Tpast}{base position matrix of positions' probability}

\item{p}{acceptance probabilities}
}
\value{
the matrix with the updated positions' probability
}
\description{
Transition of positions for a next organ
}
