% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R, R/get-center.R
\name{center}
\alias{center}
\alias{print.center}
\alias{get_center.center}
\alias{get_region.center}
\alias{get_p_accept.center}
\alias{get_offered.center}
\alias{get_state.center}
\title{Constructor for object of class \code{center}}
\usage{
center(name, region, offered = 0L, p_accept = NA, state = "italy")

\method{print}{center}(x, ...)

\method{get_center}{center}(x, ...)

\method{get_region}{center}(x, ...)

\method{get_p_accept}{center}(x, ...)

\method{get_offered}{center}(x, ...)

\method{get_state}{center}(x, ...)
}
\arguments{
\item{name}{[chr] the name of the (city) transplant center.}

\item{region}{[chr] The name of the (Italian) region in which the center
inherits (see \code{\link{regions}}).}

\item{offered}{[dbl] a single number between 0 and 1 (possibly included)
representing the number of organ offered from the center in the
previous period of reference. Default is 0.}

\item{p_accept}{[dbl] a single number between 0 and 1 (possibly included)
representing the probability to accept an offered organ. Default is
\code{NULL} meaning that the probability should be considered at region
level and not at center level.}

\item{state}{[chr] a name of available state, i.e. included in
\code{\link{names}(\link{regions})} (note: both lowercase and Titlecase are
fine for write states, i.e. "Italy" works as well as "italy").}

\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class \code{\link{center}}.
}
\description{
The function \code{\link{center}()} is the constructor function for the
homonymous class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: nice (and coloured, if supported) print method.

\item \code{get_center}: wrapper function to access to the center's name.

\item \code{get_region}: wrapper function to access to the detail "region".

\item \code{get_p_accept}: wrapper function to access to the detail "p_accept".

\item \code{get_offered}: wrapper function to access to the detail "offered".

\item \code{get_state}: wrapper function to access to the detail "state".
}}

\examples{
center('Padova', 'Veneto', 10, 0.8)
center('Padova', 'Veneto', 10, 0)
center('Padova', 'Veneto', 10)
center('Padova', 'Veneto', 10, 0.8, 'Italy')

padova <- center('Padova', 'Veneto')
get_center(padova) # Padova
get_region(padova) # veneto
get_p_accept(padova)
get_offered(padova)
get_state(padova) # italy
}
