% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-generics.R
\name{get_region}
\alias{get_region}
\title{Get the region name}
\usage{
get_region(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The region name for the object \code{x}.
}
\description{
This is a generic function aimed to access to the region name of an
object
}
