% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-center.R, R/region.R
\name{get_center.set_centers}
\alias{get_center.set_centers}
\alias{set_centers}
\alias{print.set_centers}
\title{List of centers creator}
\usage{
\method{get_center}{set_centers}(x, ...)

set_centers(...)

\method{print}{set_centers}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{a set of \code{\link{center}}}
}
\value{
The list of \code{\link{center}}s provided.
}
\description{
Function used to prepare and check list of centers to pass to
\code{centers} as needed to create \code{\link{region}} objects.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_center}: wrapper function to access to the centers' names.

\item \code{print}: nice (and coloured, if supported) print method.
}}

\examples{
pavia  <- center('Pavia',  'Lombardia')
milano <- center('Milano', 'Lombardia')
get_center(set_centers(pavia, milano)) # Pavia, Milano
pavia  <- center('Pavia',  'Lombardia')
milano <- center('Milano', 'Lombardia')

set_centers(pavia, milano)
}
