% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-set_macroregions.R, R/macroarea.R
\name{get_all_region.set_macroregions}
\alias{get_all_region.set_macroregions}
\alias{get_all_macroregion.set_macroregions}
\alias{set_macroregions}
\alias{print.set_macroregions}
\title{List of regions/macroregion creator}
\usage{
\method{get_all_region}{set_macroregions}(x, ...)

\method{get_all_macroregion}{set_macroregions}(x, ...)

set_macroregions(...)

\method{print}{set_macroregions}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{a set of \code{\link{region}}s and/or
\code{\link{macroregion}}}
}
\value{
The list of \code{\link{region}}s and \code{\link{macroregion}}s
        provided.
}
\description{
Function used to prepare and check list of \code{\link{region}}s and/or
\code{\link{macroregion}}s to pass to \code{macroregions} as needed to
create \code{\link{macroarea}} objects.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_all_region}: wrapper function to access to the all the
names of reagions appear in the `set_macroregion`,
recursively.

\item \code{get_all_macroregion}: wrapper function to access to the all the
names of macroregions (regions included).

\item \code{print}: nice (and coloured, if supported) print method.
}}

\examples{
pavia  <- center('Pavia',  'Lombardia')
milano <- center('Milano', 'Lombardia')
lombardia <- region(set_centers(pavia, milano))

padova <- center('Padova', 'Veneto', 8, 0.7)
veneto <- region(set_centers(padova))

nitp <- macroregion('NITp', set_regions(lombardia, veneto))

torino   <- center('Torino', 'Piemonte', 7, 0.6)
piemonte <- region(set_centers(torino))

get_all_region(set_macroregions(piemonte, nitp))
get_all_macroregion(set_macroregions(piemonte, nitp))
pavia  <- center('Pavia',  'Lombardia')
milano <- center('Milano', 'Lombardia')
lombardia <- region(set_centers(pavia, milano))

padova <- center('Padova', 'Veneto', 8, 0.7)
veneto <- region(set_centers(padova))

nitp <- macroregion('NITp', set_regions(lombardia, veneto))

torino   <- center('Torino', 'Piemonte', 7, 0.6)
piemonte <- region(set_centers(torino))

set_macroregions(piemonte, nitp)
}
